+++ packages/javascript/bh-shared-ui/src/views/TierManagement/Save/SelectorForm/index.tsx
@@ -127,7 +127,7 @@ const SelectorForm: FC = () => {
         <FormProvider {...formMethods}>
             <form
                 onSubmit={formMethods.handleSubmit(onSubmit)}
-                className='flex gap-6 mt-6 w-full justify-between pointer-events-auto'>
+                className='flex gap-6 mt-6 w-full max-w-[120rem] justify-between pointer-events-auto'>
                 <BasicInfo setSelectorType={setSelectorType} selectorType={selectorType} />
                 <SeedSelection
                     selectorType={selectorType}
